refactor array window slicing
