+++ crates/rulemorph/src/transform/operators/array/sequence/window/slicing/chunk.rs
@@ -1,0 +1,56 @@
+use super::*;
+
+pub(in crate::transform::operators) fn eval_array_chunk(
+    args: &[Expr],
+    injected: Option<&EvalValue>,
+    record: &JsonValue,
+    context: Option<&JsonValue>,
+    out: &JsonValue,
+    base_path: &str,
+    locals: Option<&EvalLocals<'_>>,
+) -> Result<EvalValue, TransformError> {
+    let total_len = args_len(args, injected);
+    if total_len != 2 {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr.args must contain exactly two items",
+        )
+        .with_path(format!("{}.args", base_path)));
+    }
+
+    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
+    let size_path = format!("{}.args[1]", base_path);
+    let size_value =
+        match eval_arg_value_at(1, args, injected, record, context, out, base_path, locals)? {
+            None => return Ok(EvalValue::Missing),
+            Some(value) => value,
+        };
+    if size_value.is_null() {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr arg must not be null",
+        )
+        .with_path(size_path));
+    }
+    let size = value_to_i64(&size_value, &size_path, "size must be a positive integer")?;
+    if size <= 0 {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "size must be a positive integer",
+        )
+        .with_path(size_path));
+    }
+    let size = usize::try_from(size).map_err(|_| {
+        TransformError::new(TransformErrorKind::ExprError, "size is too large").with_path(size_path)
+    })?;
+
+    let mut chunks = Vec::new();
+    let mut index = 0;
+    while index < array.len() {
+        let end = (index + size).min(array.len());
+        chunks.push(JsonValue::Array(array[index..end].to_vec()));
+        index = end;
+    }
+
+    Ok(EvalValue::Value(JsonValue::Array(chunks)))
+}
