+++ crates/rulemorph/src/transform/operators/array/sequence/window/slicing/slice.rs
@@ -1,0 +1,70 @@
+use super::*;
+
+pub(in crate::transform::operators) fn eval_array_slice(
+    args: &[Expr],
+    injected: Option<&EvalValue>,
+    record: &JsonValue,
+    context: Option<&JsonValue>,
+    out: &JsonValue,
+    base_path: &str,
+    locals: Option<&EvalLocals<'_>>,
+) -> Result<EvalValue, TransformError> {
+    let total_len = args_len(args, injected);
+    if !(2..=3).contains(&total_len) {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr.args must contain two or three items",
+        )
+        .with_path(format!("{}.args", base_path)));
+    }
+
+    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
+    let len = array.len() as i64;
+
+    let start_path = format!("{}.args[1]", base_path);
+    let start_value =
+        match eval_arg_value_at(1, args, injected, record, context, out, base_path, locals)? {
+            None => return Ok(EvalValue::Missing),
+            Some(value) => value,
+        };
+    if start_value.is_null() {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr arg must not be null",
+        )
+        .with_path(start_path));
+    }
+    let start = value_to_i64(&start_value, &start_path, "start must be an integer")?;
+
+    let end = if total_len == 3 {
+        let end_path = format!("{}.args[2]", base_path);
+        let end_value =
+            match eval_arg_value_at(2, args, injected, record, context, out, base_path, locals)? {
+                None => return Ok(EvalValue::Missing),
+                Some(value) => value,
+            };
+        if end_value.is_null() {
+            return Err(TransformError::new(
+                TransformErrorKind::ExprError,
+                "expr arg must not be null",
+            )
+            .with_path(end_path));
+        }
+        value_to_i64(&end_value, &end_path, "end must be an integer")?
+    } else {
+        len
+    };
+
+    let mut start_index = if start < 0 { len + start } else { start };
+    let mut end_index = if end < 0 { len + end } else { end };
+    start_index = start_index.clamp(0, len);
+    end_index = end_index.clamp(0, len);
+
+    let results = if end_index <= start_index {
+        Vec::new()
+    } else {
+        array[start_index as usize..end_index as usize].to_vec()
+    };
+
+    Ok(EvalValue::Value(JsonValue::Array(results)))
+}
