+++ crates/rulemorph/src/transform/operators/array/sequence/window/slicing/take_drop.rs
@@ -1,0 +1,105 @@
+use super::*;
+
+pub(in crate::transform::operators) fn eval_array_take(
+    args: &[Expr],
+    injected: Option<&EvalValue>,
+    record: &JsonValue,
+    context: Option<&JsonValue>,
+    out: &JsonValue,
+    base_path: &str,
+    locals: Option<&EvalLocals<'_>>,
+) -> Result<EvalValue, TransformError> {
+    let total_len = args_len(args, injected);
+    if total_len != 2 {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr.args must contain exactly two items",
+        )
+        .with_path(format!("{}.args", base_path)));
+    }
+
+    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
+    let count_path = format!("{}.args[1]", base_path);
+    let count_value =
+        match eval_arg_value_at(1, args, injected, record, context, out, base_path, locals)? {
+            None => return Ok(EvalValue::Missing),
+            Some(value) => value,
+        };
+    if count_value.is_null() {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr arg must not be null",
+        )
+        .with_path(count_path));
+    }
+    let count = value_to_i64(&count_value, &count_path, "count must be an integer")?;
+
+    let len = array.len() as i64;
+    let results = if count >= 0 {
+        let take_count = count.min(len).max(0) as usize;
+        array[..take_count].to_vec()
+    } else {
+        let abs_count = if count == i64::MIN {
+            (i64::MAX as u64) + 1
+        } else {
+            (-count) as u64
+        };
+        let take_count = abs_count.min(array.len() as u64) as usize;
+        let start = array.len().saturating_sub(take_count);
+        array[start..].to_vec()
+    };
+
+    Ok(EvalValue::Value(JsonValue::Array(results)))
+}
+
+pub(in crate::transform::operators) fn eval_array_drop(
+    args: &[Expr],
+    injected: Option<&EvalValue>,
+    record: &JsonValue,
+    context: Option<&JsonValue>,
+    out: &JsonValue,
+    base_path: &str,
+    locals: Option<&EvalLocals<'_>>,
+) -> Result<EvalValue, TransformError> {
+    let total_len = args_len(args, injected);
+    if total_len != 2 {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr.args must contain exactly two items",
+        )
+        .with_path(format!("{}.args", base_path)));
+    }
+
+    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
+    let count_path = format!("{}.args[1]", base_path);
+    let count_value =
+        match eval_arg_value_at(1, args, injected, record, context, out, base_path, locals)? {
+            None => return Ok(EvalValue::Missing),
+            Some(value) => value,
+        };
+    if count_value.is_null() {
+        return Err(TransformError::new(
+            TransformErrorKind::ExprError,
+            "expr arg must not be null",
+        )
+        .with_path(count_path));
+    }
+    let count = value_to_i64(&count_value, &count_path, "count must be an integer")?;
+
+    let len = array.len() as i64;
+    let results = if count >= 0 {
+        let drop_count = count.min(len).max(0) as usize;
+        array[drop_count..].to_vec()
+    } else {
+        let abs_count = if count == i64::MIN {
+            (i64::MAX as u64) + 1
+        } else {
+            (-count) as u64
+        };
+        let drop_count = abs_count.min(array.len() as u64) as usize;
+        let end = array.len().saturating_sub(drop_count);
+        array[..end].to_vec()
+    };
+
+    Ok(EvalValue::Value(JsonValue::Array(results)))
+}
